+++ ml_tools/schema.py
@@ -1,4 +1,7 @@
-from ._core._schema import FeatureSchema
+from ._core._schema import (
+    FeatureSchema,
+    info
+)
 
 __all__ = [
     "FeatureSchema"