+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [18.3.1] 2025-11-28
+
+### Fixed
+
+- Minor public API cleaning.
+
+- ML_utilities:
+    - `ArtifactFinder`, correctly exposes the model architecture path.
+
 ## [18.3.0] 2025-11-27
 
 ### Added